refactor: move the SSO refresh helpers into core
The keyring provider imported refresh_access_token from cli.saml,
so core depended on cli at import time. Move refresh_access_token
and raise_for_api_error into core/sso.py. cli/saml.py re-exports
them for its existing callers.

The deferred provider import in chain.py stays. decorators.py
imports chain on every CLI start, so a module-level provider
import would load requests and cloudsmith_api at startup and fail
test_cli_import_does_not_load_heavy_modules.
