+++ cloudsmith_cli/cli/saml.py
@@ -2,7 +2,15 @@ from urllib.parse import urlencode
 
 import requests
 
-from ..core.api.exceptions import ApiException
+from ..core.sso import raise_for_api_error, refresh_access_token
+
+__all__ = [
+    "create_configured_session",
+    "exchange_2fa_token",
+    "get_idp_url",
+    "raise_for_api_error",
+    "refresh_access_token",
+]
 
 
 def create_configured_session(opts):
@@ -24,28 +32,6 @@ def create_configured_session(opts):
         session.headers.update(opts.api_headers)
 
     return session
-
-
-def raise_for_api_error(response):
-    """Raise :class:`ApiException` if *response* failed, keeping the API's detail.
-
-    Without the detail the exception renders as bare status text, so a caller
-    reporting it tells the user that something failed but never what.
-    """
-    try:
-        response.raise_for_status()
-    except requests.RequestException as exc:
-        try:
-            body = exc.response.json()
-        except ValueError:
-            body = None
-
-        raise ApiException(
-            response.status_code,
-            detail=body.get("detail") if isinstance(body, dict) else None,
-            headers=exc.response.headers,
-            body=exc.response.content,
-        )
 
 
 def get_idp_url(api_host, owner, session):
@@ -80,27 +66,5 @@ def exchange_2fa_token(api_host, two_factor_token, totp_token, session):
     exchange_data = exchange_response.json()
     access_token = exchange_data.get("access_token")
     refresh_token = exchange_data.get("refresh_token")
-
-    return (access_token, refresh_token)
-
-
-def refresh_access_token(api_host, access_token, refresh_token, session):
-    data = {"refresh_token": refresh_token}
-    url = f"{api_host}/user/refresh-token/"
-
-    headers = {"Authorization": f"Bearer {access_token}"}
-
-    response = session.post(
-        url,
-        data=data,
-        headers=headers,
-        timeout=30,
-    )
-
-    raise_for_api_error(response)
-
-    response_data = response.json()
-    access_token = response_data.get("access_token")
-    refresh_token = response_data.get("refresh_token")
 
     return (access_token, refresh_token)
