+++ cloudsmith_cli/core/credentials/providers/keyring_provider.py
@@ -4,9 +4,9 @@ from __future__ import annotations
 
 import logging
 
-from ....cli.saml import refresh_access_token
 from ....core import keyring
 from ...api.exceptions import ApiException
+from ...sso import refresh_access_token
 from ..models import CredentialContext, CredentialResult
 from ..provider import CredentialProvider
